+++ src/app/static/src/app/store/modelCalibrate/actions.ts
@@ -82,7 +82,7 @@ export const actions: ActionTree<ModelCalibrateState, RootState> & ModelCalibrat
                 if (switches.modelCalibratePlot) {
                     dispatch("getCalibratePlot");
                 }
-                dispatch("getComparisonPlot");
+                await dispatch("getComparisonPlot");
             }
         }
         commit(ModelCalibrateMutation.Ready);