+++ NEWS.md
@@ -1,6 +1,18 @@
-# hint 2.23.0 
+# hint 2.26.0 
 
 * Remove fixed disaggregate value from xAxis options in vue-charts 
+
+# hint 2.25.0
+
+* Login to data exploration throws an error on redirect
+
+# hint 2.24.0
+
+* Barchart uncertainty ranges (error bars) are greyed out on first rendering
+
+# hint 2.23.0
+
+* Implement no data message in comparison chart
 
 # hint 2.22.0